moved compress to utils
